App refactor: Introduce AppContextProvider
Improve App module's read- and compose-ability by moving AppContext
concerns to a separate provider module.
